Create replacer function for json response
Added replacer so that the dates could only be returned as YYYY-MM-DD,
otherwise json.stringify will convert the date to full ISO string
with Z, denoting UTC time.
This would cause all the dates to be converted incorrectly based on the
user's time zone; I only need the date, not the time or timezone.