Restructure file structure again
When adding new auth feature, it felt more natural to set it up in its
own separate folder. I then transformed the calendar components in its
own feature to follow.

Now this structure should be quite close to the bulletproof-react model,
except for a few differences such as everything being accessed via
an index.ts, as well as Layouts. I will most likely add those in a
future project.

I also tried to import the svg as components, but decided a/g it, while
also adding aria labels to buttons that only have icons.

I also renamed the Toast component to be a Notification, along with
changing the store, for when I use it more generally as a notification
for actions, errors, etc.