Refactor routes to use headers for uid
It seems a better method to use a header with every request in order
to retrieve the correct plants. Also added a bit more verification that
the uid is in fact a string, as well as adding uid verification to
the delete plant route, otherwise no id would be needed to delete any
plant.